+++ packages/kit/src/core/make_package/index.js
@@ -1,8 +1,9 @@
 import * as fs from 'fs';
+import globrex from 'globrex';
+import { createRequire } from 'module';
 import * as path from 'path';
 import { preprocess } from 'svelte/compiler';
-import globrex from 'globrex';
-import { mkdirp, rimraf } from '../filesystem';
+import { mkdirp, rimraf } from '../filesystem/index.js';
 
 /**
  * @param {import('types/config').ValidatedConfig} config
@@ -11,8 +12,14 @@ import { mkdirp, rimraf } from '../filesystem';
 export async function make_package(config, cwd = process.cwd()) {
 	rimraf(path.join(cwd, config.kit.package.dir));
 
+	// Generate type definitions first so hand-written types can overwrite generated ones
+	await emit_dts(config);
+
 	const files_filter = create_filter(config.kit.package.files);
-	const exports_filter = create_filter(config.kit.package.exports);
+	const exports_filter = create_filter({
+		...config.kit.package.exports,
+		exclude: [...config.kit.package.exports.exclude, '*.d.ts']
+	});
 
 	const files = walk(config.kit.files.lib);
 
@@ -57,15 +64,24 @@ export async function make_package(config, cwd = process.cwd()) {
 
 		if (svelte_ext) {
 			// it's a Svelte component
-			// TODO how to emit types?
 			out_file = file.slice(0, -svelte_ext.length) + '.svelte';
 			out_contents = config.preprocess
 				? (await preprocess(source, config.preprocess, { filename })).code
 				: source;
-		} else if (ext === '.ts' && !file.endsWith('.d.ts')) {
-			// TODO transpile TS file and emit types
-			// also, we want to emit types from JSDoc annotations in .js files
-			throw new Error('svelte-kit package does not yet support TypeScript');
+		} else if (ext === '.ts' && file.endsWith('.d.ts')) {
+			// TypeScript's declaration emit won't copy over the d.ts files, so we do it here
+			out_file = file;
+			out_contents = source;
+			if (fs.existsSync(path.join(cwd, config.kit.package.dir, out_file))) {
+				console.warn(
+					'Found already existing file from d.ts generation for ' +
+						out_file +
+						'. This file will be overwritten.'
+				);
+			}
+		} else if (ext === '.ts') {
+			out_file = file.slice(0, -'.ts'.length) + '.js';
+			out_contents = await transpile_ts(filename, source);
 		} else {
 			out_file = file;
 			out_contents = source;
@@ -96,6 +112,59 @@ export async function make_package(config, cwd = process.cwd()) {
 	if (fs.existsSync(project_readme) && !fs.existsSync(package_readme)) {
 		fs.copyFileSync(project_readme, package_readme);
 	}
+}
+
+/**
+ * @param {string} filename
+ * @param {string} source
+ */
+async function transpile_ts(filename, source) {
+	const ts = await try_load_ts();
+	return ts.transpileModule(source, {
+		compilerOptions: load_tsconfig(filename, ts),
+		fileName: filename
+	}).outputText;
+}
+
+async function try_load_ts() {
+	try {
+		return (await import('typescript')).default;
+	} catch (e) {
+		throw new Error(
+			'You need to install TypeScript if you want to transpile TypeScript files and/or generate type definitions'
+		);
+	}
+}
+
+/**
+ * @param {string} filename
+ * @param {import('typescript')} ts
+ */
+function load_tsconfig(filename, ts) {
+	const filedir = path.dirname(filename);
+	const tsconfig_filename = ts.findConfigFile(filedir, ts.sys.fileExists);
+
+	if (!tsconfig_filename) {
+		throw new Error('Failed to locate tsconfig or jsconfig');
+	}
+
+	const { error, config } = ts.readConfigFile(tsconfig_filename, ts.sys.readFile);
+
+	if (error) {
+		throw new Error('Malformed tsconfig\n' + JSON.stringify(error, null, 2));
+	}
+
+	// Do this so TS will not search for initial files which might take a while
+	config.include = [];
+	config.files = [];
+	const { options } = ts.parseJsonConfigFileContent(
+		config,
+		ts.sys,
+		path.dirname(tsconfig_filename),
+		{ sourceMap: false },
+		tsconfig_filename
+	);
+	return options;
 }
 
 /**
@@ -147,4 +216,31 @@ function walk(cwd) {
 function write(file, contents) {
 	mkdirp(path.dirname(file));
 	fs.writeFileSync(file, contents);
+}
+
+/**
+ * @param {import('types/config').ValidatedConfig} config
+ */
+export async function emit_dts(config) {
+	const require = createRequire(import.meta.url);
+	const emit = await try_load_svelte2tsx();
+	emit({
+		libRoot: config.kit.files.lib,
+		svelteShimsPath: require.resolve('svelte2tsx/svelte-shims.d.ts'),
+		declarationDir: config.kit.package.dir
+	});
+}
+
+async function try_load_svelte2tsx() {
+	try {
+		const svelte2tsx = (await import('svelte2tsx')).emitDts;
+		if (!svelte2tsx) {
+			throw new Error('Old svelte2tsx version');
+		}
+		return svelte2tsx;
+	} catch (e) {
+		throw new Error(
+			'You need to install svelte2tsx >=0.4.1 if you want to generate type definitions'
+		);
+	}
 }