+++ packages/kit/src/core/make_package/test/fixtures/typescript/expected/Test2.svelte
@@ -1,0 +1,3 @@
+<script lang="ts">
+    export let foo;
+</script>