+++ packages/kit/src/runtime/server/index.js
@@ -33,11 +33,13 @@ export async function respond(incoming, options, state = {}) {
 	}
 
 	try {
+		const headers = lowercase_keys(incoming.headers);
+
 		return await options.hooks.handle({
 			request: {
 				...incoming,
-				headers: lowercase_keys(incoming.headers),
-				body: parse_body(incoming),
+				headers,
+				body: parse_body(incoming.rawBody, headers),
 				params: null,
 				locals: {}
 			},